requirements.txt
Django==5.1
crispy-bootstrap5
django-crispy-forms
Pillow
django-filter
djangorestframework
openpyxl>=3.1.0
weasyprint>=60.0
